+++ projects/swimlane/ngx-ui/src/lib/components/calendar/calendar.module.ts
@@ -1,6 +1,5 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule } from '@angular/forms';
 import { MomentModule } from 'ngx-moment';
 
@@ -10,6 +9,6 @@ import { PipesModule } from '../../pipes/pipes.module';
 @NgModule({
   declarations: [CalendarComponent],
   exports: [CalendarComponent],
-  imports: [CommonModule, FormsModule, MomentModule, FlexLayoutModule, PipesModule]
+  imports: [CommonModule, FormsModule, MomentModule, PipesModule]
 })
 export class CalendarModule {}